+++ src/constant/index.ts
@@ -1,3 +1,5 @@
+import type { FeeConfig } from "@/types/tx.types";
+
 export const POWER_BY = "aurowallet.com";
 
 // ============ Coin Config ============
@@ -121,3 +123,18 @@ export const TokenLaunch =
 
 export const TRANSACTION_FEE = 0.1001;
 export const ZEKO_FEE_INTERVAL_TIME = 5;
+
+// ============ Fee Config ============
+
+
+export const DEFAULT_FEE_CONFIG: FeeConfig = {
+  version: 1,
+  transactionFee: {
+    slow: "0.0011",
+    medium: "0.0101",
+    fast: "0.2001",
+  },
+  feeCap: "10",
+  speedUpBuffer: "0.5",
+  zkAppAccountUpdateFee: "0.002",
+};